[refactor] Remove a shellcheck ignore directive
Refactor the way the `ionice` command is added to the backup command, in a way which passes shellcheck tests, removing the need for a shellcheck "ignore" directive.
It is important that the `ionice` command variable isn't quoted, so as not to pass an empty argument when it is null or empty.